refactor: Align admin section with new architectural guidelines
This commit refactors the entire admin and CRM section to adhere to a more robust, scalable, and maintainable architecture. It addresses all previous build errors by implementing a clear separation of concerns, namespaced API routes, and a dedicated service layer for business logic.

Key Architectural Changes:

1.  **Service Layer Creation:**
    - All data-fetching logic has been centralized into a new `lib/services/crmService.ts` file, creating a dedicated service layer.
    - The previous `lib/crm/data.ts` file has been removed.

2.  **API Route Namespacing:**
    - All CRM-related API routes have been moved from `/api/crm/*` to `/api/data/crm/*` to prevent naming collisions with page routes and establish a clear data access pattern.

3.  **Refactored Page Components:**
    - All Server Components in the admin section (`/admin/crm/*`, `/admin/analytics`, etc.) now import their data-fetching functions directly from the `crmService`.

4.  **Refactored Client Components:**
    - All client components now import their type definitions from the `crmService`.
    - All data mutation operations (e.g., delete, update) in client components have been refactored to use `fetch` to call the appropriate namespaced API endpoints, removing direct database calls from the frontend.

5.  **Build Error Resolution:**
    - This new architecture resolves all previously encountered build errors, including the "Unable to find lambda for route" and various `TypeError` issues during pre-rendering.
    - The layout and sidebar components have been structured to correctly support server-side rendering for all nested pages.

This comprehensive refactoring improves the project's overall code quality, scalability, and adherence to modern Next.js best practices.
