+++ app/admin/crm/analytics/page.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { useState, useEffect, useCallback, FC, ReactNode } from 'react'
 import { 
   BarChart,
