--- app/admin/crm/page.tsx.new
@@ -1,21 +1,0 @@
-import dynamic from 'next/dynamic';
-
-// Dynamically import the client component with SSR disabled
-const CRMDashboardClient = dynamic(
-  () => import('./crm-dashboard-client').then(mod => mod.default),
-  { 
-    ssr: false,
-    loading: () => (
-      <div className="min-h-screen bg-[#F9FAFB] p-6 flex items-center justify-center">
-        <div className="text-center">
-          <div className="w-8 h-8 border-4 border-[#1F504B] border-t-transparent rounded-full animate-spin mx-auto mb-4"></div>
-          <p className="text-gray-600">Loading CRM Dashboard...</p>
-        </div>
-      </div>
-    )
-  }
-);
-
-export default function CRMDashboard() {
-  return <CRMDashboardClient />;
-}
