+++ dev.log
@@ -1,3 +1,3 @@
 
-> my-v0-project@0.1.0 dev
+> my-v0-project@0.1.0 dev /app
 > next dev
